Osztályok átalakítása, funkciók kiszervezése
