+++ src/main/java/hu/oe/nik/szfmv/automatedcar/bus/VirtualFunctionBus.java
@@ -1,7 +1,7 @@
 package hu.oe.nik.szfmv.automatedcar.bus;
 
+import hu.oe.nik.szfmv.automatedcar.bus.packets.inputposition.ReadOnlyInputPositionPacket;
 import hu.oe.nik.szfmv.automatedcar.bus.packets.powertrain.ReadOnlyPowertrainPacket;
-import hu.oe.nik.szfmv.automatedcar.bus.packets.sample.ReadOnlySamplePacket;
 import hu.oe.nik.szfmv.automatedcar.systemcomponents.SystemComponent;
 
 import java.util.ArrayList;
@@ -15,7 +15,7 @@ import java.util.List;
  */
 public class VirtualFunctionBus {
 
-    public ReadOnlySamplePacket samplePacket;
+    public ReadOnlyInputPositionPacket readOnlyInputPositionPacket;
     public ReadOnlyPowertrainPacket powertrainPacket;
 
     private List<SystemComponent> components = new ArrayList<>();
