+++ src/main/java/hu/oe/nik/szfmv/automatedcar/systemcomponents/Driver.java
@@ -1,20 +1,23 @@
 package hu.oe.nik.szfmv.automatedcar.systemcomponents;
 
 import hu.oe.nik.szfmv.automatedcar.bus.VirtualFunctionBus;
-import hu.oe.nik.szfmv.automatedcar.bus.packets.sample.SamplePacket;
+import hu.oe.nik.szfmv.automatedcar.bus.packets.inputposition.InputPositionPacket;
 
 public class Driver extends SystemComponent {
 
-    private final SamplePacket samplePacket;
+    private final InputPositionPacket inputPositionPacket;
 
+    /**
+     * @param virtualFunctionBus Create a Driver with Virtual Function Bus
+     */
     public Driver(VirtualFunctionBus virtualFunctionBus) {
         super(virtualFunctionBus);
-        samplePacket = new SamplePacket();
-        virtualFunctionBus.samplePacket = samplePacket;
+        inputPositionPacket = new InputPositionPacket();
+        virtualFunctionBus.readOnlyInputPositionPacket = inputPositionPacket;
     }
 
     @Override
     public void loop() {
-        samplePacket.setGaspedalPosition(5);
+        inputPositionPacket.setGaspedalPosition(5);
     }
 }
